refactor: separate logic to dif files
